requirements.txt
fastapi[standard]
uvicorn
python-dotenv
pytest
supabase
black
isort